The list of available studies is now displayed with a fragment
Because:
- This is the next step to restructure the app

The different contents of the app are displayed with fragments and a central activity.
